backend/requirements.txt
transformers
torch
PyMuPDF
python-docx
python-pptx
pytesseract
pillow
camelot-py[cv]
pandas
opencv-python
tensorflow
glob2
matplotlib
nltk